src/backend/post_service/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
alembic==1.13.1
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
orjson==3.9.15
protobuf==4.25.3
psycopg2==2.9.9
pydantic==2.6.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pytest==8.1.1
pytest-asyncio==0.23.6
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
ujson==5.9.0
uvicorn==0.28.0
watchfiles==0.21.0
websockets==12.0
gunicorn==21.2.0
greenlet==3.0.3
uvicorn==0.28.0
